Change Akka to Pekko; bump Scala to 3.3.0 (#13)
Pekko has a 1.0.0 release now, so we can safely use it. Akka's OSS
version will not really be updated, so that's the way to go. Pekko for
Scala 3 requires (I think) Scala 3.3.0, so I've updated it as well.

aand some minor bumps to other dependencies (patches only)

Everything seems to work fine.