Add RDF4J parser/writer for Jelly (#22)
Added interop code to make it possible to use Jelly as any other
language supported in Rio. For serialization, all config options for
Jelly are available through WriterConfig. For deserialization, the
parser auto-detects the file format and adjusts accordingly.

Added tests cross-checking this code with the Jena code, it seems to all
work well.