Rework stream frame size limits (#41)
The previous system required the user to use a byte size limit on the
stream frames, which did not make sense in all cases (e.g., some grouped
streaming use cases).

This commit introduces the SizeLimiter trait that can be used to define
the policy for splitting frames when they get too big. The policy can be
disabled entirely in grouped streaming formulations, but is mandatory in
flat streaming.