Implement RDF-STaX in Jelly (#53)
Issue #51, #52

Implemented:

- Physical & logical stream types in Jelly, where logical are directly
from RDF-STaX
  - Logical types are implemented as an enum
- Type numbers were picked such that the most common types (1, 2, 3, 4)
are low, making the varint in the protobuf output smaller.
- Type numbers signify taxonomical relations in RDF-STaX. 14 is a child
of 4, and 114 is a child of 14. This makes the system forward compatible
– current consumers will be able to interpret future subtypes as their
parent types.
- Support for logical types in streaming encoders/decoders
- Automatic checks for logical type compatibility and logical-physical
consistency (if a given stream is possible, can be consumed, etc.)
- Emitting RDF-based annotations (metadata) about the logical type of
the stream, using the RDF-STaX ontology. This uses the existing
facilities of decoder converters.