Implement max supported stream options in the decoder (#94)
The decoder should refuse to process streams that use options not matching the user's expectations. This includes: overly large lookup tables (can lead to OOM – DoS attack vector), unwanted RDF-star, unwanted generalized statements, limiting the protocol version to something lower than what's supported by the library.

This was implemented in core, along with a refactor of how stream option checks work in general. The stream module (DecoderFlow) was also refactored to allow passing supported options to the decoder. At the same time, the ugly boolean strict parameter was removed there and replaced with the more meaningful method variants. Lastly, this was implemented in the RDF4J/Jena parsers.

There are extensive tests and examples for all of this.