Build plugins for Jena and RDF4J (#97)
These plugins should be simply included in Jena's/RDF4J's classpath to enable Jelly support. The plugins do not include Jena/RDF4J themselves, as this is not needed – only the protobuf stuff, Scala standard library, and our code.

The plugins will be published to GH releases – let's see if it works.