Add fake Scala 2 builds
This is a blatant misuse of sbt's facilities to make it possible to publish a Scala 3 library compiled using the Scala 3 compiler, but with two different sets of dependencies: for Scala 2 and 3. This *SHOULD* allow others to use Jelly in Scala 2 projects without worrying about conflicting transitive dependencies. I hope.