Experiment: Use ByteStrings instead of Strings in protos
I've noticed in profiling that a lot of the time is spent on computing the serialized sizes of strings... which is because the protobuf lib has to manually count all the code points because Java Strings are horrible – long story.

Let's see if this helps anything. If not, I will revert it.

WHY ARE THERE NO UTF-8 STRINGS IN JAVA? WHY???