Fix handling of Jena's default graph in parser
See: https://github.com/apache/jena/issues/2578

We used to emit "null" as the default graph, which (I think) is incorrect. We should be outputting Quad.defaultGraphIRI, which is correctly handled by the rest of Jena.

I've added a test for the parser that should pass now and a test for the serializer if it covers all types of default graphs in Jena.