Experiment: unify term classes in proto code (#117)
Use scalameta to completely remove the generated oneofs in RdfTriple and RdfQuad, instead using a single class hierarchy for RDF terms, defined in jelly-core.

This does simplify the class hierarchy a lot. It may help performance especially in the encoder because we now have no need to pass around the annoying term adapter implicits. Instead, we can construct and deconstruct terms directly.