Experiment: remove instanceof in term handling (#119)
* Experiment: remove instanceof in term handling

Continuation of #118

The previous PR did improve deserialization speed significantly, but it made serialization slightly slower, which was not expected. I assume that might be due to the large number of instanceof checks and class casts involved. I've replaced this with a hybrid of the previous solution, where only for blank nodes we use a custom wrapper class (because String is final), and for the rest (e.g., RdfIri), we simply extend their interfaces to include the checker and accessor methods that we need from them.

Let's see if this helps. If not, I will revert this together with #118.

* Remove one remaining asInstanceOf

* Sorry, one more