Encoder: maybe speed up the node cache
The LRU policy there is probably way overkill, we can do this much more easily with a dead-simple hashed cache. This implementation is copied over from Jena, let's see if this works.

This will result in more evictions than needed, but the lookups will be way, way faster...

As always, I will finish the tests and docstrings if this works. If not, I will revert it.