Experiment: unwrap RdfStreamRow contents (#198)
Unwrap RdfStreamRow contents

This is the same story as with #118, just a bit more straightforward.

This will save us one allocation per stream row in serialization AND deserialization, and also reduce the overhead of going through the ugly if-else chains. Now the JVM should use the tableswitch instruction which is way, way faster.

Let's see if it helps. If not, I will revert the whole thing. If yes, I will complete the docstrings.