Add an IRI cache to NodeDecoder (#201)
* Add an IRI cache to NodeDecoder

Also, rewrite it in Java.

As usual, if it works, I will finish the docstrings etc.

This is a very simple cache that assumes that a name entry is most likely used only with one prefix entry. Therefore, the last IRI constructed from prefix+name is cached here. There is a mechanism for dependent cache invalidation very similar to what we have in NodeEncoder (with serial numbers).

Let's see.

* Remove old code

* Really remove old code...

* Fix bug with overriding name entries