Experiment: Try to improve lookup coherence
Implement a simple check in the encoder lookup to see it's possible to withdraw new IDs from it in a more cache-friendly and wire representation-friendly manner. Maybe it will help with compression ratios and performance. Maybe. I'm not entirely convinced that the algorithm is very good. There are better policies for cache reordering, but they would be much more costly to implement.