Remove dependency on full Netty
The dependency was not needed, because grpc-netty-shaded already provides what we need: https://github.com/grpc/grpc-java/issues/10931

This also helps avoid any future vulnerabilities, because Netty is quite prone to these.

Note that this is not a breaking change, because we weren't distributing any "real" packages with this dependency.