Fix missing phys. stream type in RDF4J writer
Issue: #218

We cannot fix this one in the same way as in Jena, because in RDF4J there is no way to tell apart a triple and a quad in the default graph. Because of that, the RDF4J writer will now default to writing a QUADS stream, unless it's told otherwise.

I've updated the tests and scaladocs accordingly to reflect this.