Refactor: use a params case class for ProtoEncoder
Instead of messily updating argument lists and creating a ton of overloaded factory methods, let's just pass a case class instance that contains all parameters for the encoder.

This has two benefits. First, it will be much easier to expand the list of parameters for the ProtoEncoder. Second, handling these params will be much easier in functional code, because now we have them grouped in one tidy object.