Improve exception handling in core (#307)
This tidies up exception handling in decoders and encoders, reducing the chance that some NPE, array-out-of-bounds, class-cast, or whatever exception slips out. There would be nothing inherently *wrong* with that, as it's all unchecked exceptions anyway, but such exceptions are not very informative to users. By catching and re-throwing them, we can make it clearer why a failure has occurred.