Make AnyStatementDecoder ignore duplicate options
This was the only decoder that threw an error when encountering duplicate stream options, which was pretty inconsistent. This was technically OK with the spec ("Implementations SHOULD NOT throw an error if the stream options header is present more than once in the stream."), but was not the RECOMMENDED behavior.