Experimental Jelly-Patch implementation (#333)
* Add ProtoDecoder.ingestRowFlat

This method avoids the overhead of allocating Some[Statement] during parsing. The old .ingestRow method is kept both for back-compat and just because it's still useful.

You may want to use ingestRowFlat if you are working with Java code or want to squeeze out an extra bit of performance.

* scaladoc fix

* WIP: RDF Patch implementation

Turns out, lots of preparatory work in core will be required.

* make it compile

* nodeencoder shenanigans

* Patch row values

* more patch row values

* RdfPatchHeader

* complete core implementation

* factory & tests

* added opt handling & constants, WIP tests

* tests WIP

* finish encoder tests

* PatchDecoder WIP

* WIP decoder

* Expose NameDecoder for extensions via trait

* WIP: decoding

* fix things

* decoder implemented

* update for the new row types

* More things

* tests, decoder factory methods

* WIP: decoder tests

* decoder tests amen

* scaladoc

* new proto

* add symlink

* try to fix compilation error

* flatten AnyStatementDecoder