Add crunchy-protoc-plugin
Issue: #181

This merges in the protoc plugin code that used to live here: https://github.com/Jelly-RDF/protoc-java-fast That project will be archived.

This compiler generates highly optimized Java code that is similar to what we had in the Scala version of Jelly-JVM, after extensive metaprogramming and manual modifications.

This PR only covers adding the compiler code, without any tests or hooking it up to the build process --- this will be done in a follow-up PR.