Java rewrite: use the new proto compiler (#357)
This hooks up the new protoc plugin with the build infrastructure and adapts the existing Java code to the new proto classes.

The way this new plugin is called is a bit hacky, but I guess we can clean up build.sbt later.

There are also a few small imperfections in the code that will need to be resolved before 3.0.0, but that's all listed under #348