Implement integration tests for Java rewrite, part 1
Created the `integration-tests-java` module with tests mostly just for RDF parsing and serialization using the native APIs.

Also added the QuadEncoder interface. Added impls for Triple and QuadEncoders in jena-java and rdf4j-java.

And fixed a bunch of bugs that I've found, mostly in the Jena and RDF4J integrations.

Not done:

- Forward compat spec – this requires talking to the Proto compiler, which I'm not in the mood of doing at the moment.
- Reactive streams-based ser/des implementations in IO tests.
- Transcoding tests (they used the reactive API)
- CrossStreamingSpec
- GeneralizedRdfSpec (uses the reactive API...)
- Tests with Titanium