Fix serializing literals with xsd:string in Jena
For some reason, the Java code was looking for datatype == null, while it should be comparing by reference to XSDstring (yes, by reference, I'm 110% sure, I've even fixed a bug related to this in Jena).

This led to suboptimal compression and worse size results as compared to Jelly-JVM 2.10. With this fix, we are on exact size parity.