Optimize the Java ser/des implementations (#373)
A few deep optimizations to the new serializer / deserializer implementations, to hopefully bring it to performance parity with 2.10.x. Turns out that it's a really hard task, if you've previously spent... way too much time optimizing the Scala implementation.

This focuses on the Jena and RDF4J integrations. For the others, I've opened follow-up tasks: #376 #375