Remove redundant readTag call in tableswitch-based mergeFrom (#397)
In the default case (skipping an unknown field) we had a redundant call to readTag(). This should only be done in the fallthrough-optimized version, as in that case it is the case body that reads the tag. In the tableswitch version, the tag is read at the start of the loop body instead.