WIP: optimize encoder/decoder by removing extra objects
This should reduce the number of references we need to navigate on runtime... and also reduce the amount of objects we need to always store in L1 cache, freeing up a few cache lines for the input and output to the encoder/decoder. I'm in the middle of implementing this, so tests still fail. TODO: finish this and JMH it