Remove indirect references in encoder/decoder (#413)
Issue: #406

This removes LazyProperty and LastNodeHolder classes, which serve to simply hold references to other objects and wrap the logic for accessing them.

Unfortunately, they are also used in a hot path, and each of these objects adds 16 bytes of heap usage and possibly additional pointer math. The more significant part is the 16 bytes, because that takes up cache lines, and cache throughput is our main bottleneck.

I've simply unrolled these objects, putting the relevant access logic in the containing classes, as private methods.