Fix parsing always-empty 0-length messages
spin-off of #426, issue #381

We had trouble there with parsing non-delimited always empty messages. This was because of an unhandled case in a special version of the parsing code that's generated only for messages that don't have any fields.

I've added a regrtest for this. I've also checked this on top of #426 and it fixes the issue there.