Reduce CI log sizes for tests (#450)
CI test logs are 15239 lines at the moment – a large part of that is of course the 8 or 9k tests that we have... but also a bunch of random output from jelly-cli, some tested methods, and so on.

I added an env var to silence all output within specific code blocks. This works only with Scala's Console object, as it has proper output redirection support. For Java this is a complete unworkable mess, so the Java examples still output a bit of stuff to the console.

I also resolved a few other minor issues with tests being loud (e.g., stray debug messages).

This reduced the log size to 9739 lines. Yay.