JenaChangesCollector: speed up accessing internal collection
ArrayList is thread-safe-y, we don't need that, so we use Vector instead. I added an extension of Vector that lets us access the underlying data array directly, which should slightly speed up iteration over it.

I'm doing this because I noticed that in a serialization benchmark, iterating the ArrayList took 7% of the total benchmark time... that's a lot.