rdf4j: allow overriding ValueFactory in parser
Until now, the Jelly parser always used the SimpleValueFactory, regardless of settings. This is not great when directly parsing into a Repository/Sail, where we really should be using that repository's Value and Statement implementations to avoid extra overhead. This change makes it possible to do so.

This is also needed for our internal project.