ProtoMessage: increase OutputStream buffer size
RDF messages quite commonly contain long strings. If we are writing these to an OutputStream, Protobuf will first try to fit the string in its fixed-size buffer (4kB by default), and, if it fails, will create a wholly new buffer for encoding the string.

This is pretty wasteful. I think the correct solution would be to change this part of the Protobuf implementation to write the string in chunks fitting the buffer, but it may be hard to reconcile with other requirements that this class has.

Instead, we can simply raise the buffer size 2x to accommodate larger strings. Note that if the message is known to be smaller than 8kB, a smaller buffer will be allocated instead.