PatchEncoder: use MessageCollection for row buffering
First step towards using reusable row buffers – use the MessageCollection class to avoid copying the buffer to RdfPatchFrame. This also allows us to use a reusable frame in the Jena writer, which should also speed things up a little bit.

As a side-quest I double-checked if non-delimited Jelly data is correctly written by native writers and yeah, it is. I also removed a redundant cached size reset, as in non-delimited serialization the cached size is not used for anything.