Implement class-based dispatch for term handling (#469)
* Experimental ProtoEncoder passing state on stack

Theoretically this should allow for better inlining and less load/store operations, but I haven't seen any benefit in benchmarks (no statistically significant difference) – maybe they were too noisy, as I was running a full serializer in the benchmark.

Come back to this later... I think the idea is solid, but it needs a better (more precise) benchmark.

# Conflicts:
#	core-patch/src/main/java/eu/neverblink/jelly/core/patch/internal/PatchEncoderImpl.java

* I dug into why it was so slow – apparently we were allocating these lambdas as objects and binding them dynamically... that's no good.

This is a solution with Records – they should (?) get inlined by the JVM.

* remove unused stuff

* The previous one was even slower... the records did not get properly inlined.

This is a radical "new" take that's basically identical to the original Scala code. 2x less load/stores, a few saved slots for numbers, and less code. Let's see if it works.

* Speedier type checks

Went from 9 to 3 instructions per typecheck, OpenJDK 23, Ryzen 4

* fix

* fix2