Native parsers: use a single CodedInputStream
Very similar optimization to #467

We allocate only one CodedInputStream for the entire delimited stream.

This allows us to make a more efficient use of the buffer (having multiple messages in the buffer are fine). We also remove the need to allocate a new buffer for each frame. Finally, we can also remove the LimitedInputStream wrapper, as the parser is free to consume the entire input stream.