RDF4J: disable checking by default, allow enabling via factory setting
Unfortunately, the term validation stack in RDF4J is very heavy, and with it the Jelly parser become 2–4x slower. I think it makes more sense to disable all the checks by default.

To make it possible to enable these checks in RDF4J parser consistency tests, I also added a setting on the parser factory itself, to control whether parser instances produced by it will be checking or not.