Jena/RDF4J: increase default size of serializer lookups
We used to default to the very small name/prefix tables that reduce performance drastically when dealing with larger datasets. Meanwhile, in small datasets these larger lookups perform the same or just slightly worse than the small ones. It makes more sense to use the larger ones as the default.